refactor: extract CredentialTokenField.swift from SettingsView.swift
Extracts CredentialTokenField (NSViewRepresentable, 87 lines) +
CredentialTokenTextField (NSTextField subclass, 43 lines) from
SettingsView.swift into a new CredentialTokenField.swift.

Byte-identical move; SHA-256 93a38d4efaab85f521d13b306780bd9891741187e66eb95fa581e95da5d48686
verified against origin/main:SettingsView.swift lines 283-414.

Imports: AppKit + SwiftUI.

Reduces SettingsView.swift from 414 to 282 lines.

Tests: 667 (unchanged).

⚠️ Manual smoke check recommended before merge: verify credential (API key)
text fields in Settings still behave identically (paste, secure-entry, focus).
